text to speak now added, although not the voice i want in the end product. Works as expected for now. Also added some greeting messages.

+++ main/general_interaction/date_and_time.py
@@ -1,7 +1,7 @@
 import datetime
 import sys
 
-def get_date(date_format='long'):
+def get_date(date_format='short'):
     # Get the current date in a short or long format by filling the type parameter
     today = datetime.date.today()
